requirements.txt
-i https://pypi.org/simple
asgiref==3.2.5
beautifulsoup4==4.6.0
certifi==2019.11.28
chardet==3.0.4
django-environ==0.4.5
django-modelcluster==5.0.1
django-taggit==1.2.0
django-treebeard==4.3.1
django-webpack-loader==0.7.0
django==3.0.4
djangorestframework==3.11.0
draftjs-exporter==2.1.7
html5lib==1.0.1
idna==2.9
l18n==2018.5
pillow==6.2.2
psycopg2==2.8.4
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
unidecode==1.1.1
urllib3==1.25.8
wagtail==2.8
webencodings==0.5.1
willow==1.3
